home screen completed

+++ app/src/main/java/com/example/weather/data/model/weather/AirQuality.kt
@@ -7,7 +7,7 @@ data class AirQuality(
     @Json(name = "gb-defra-index") val gbDefraIndex: Int,
     val no2: Double,
     val o3: Double,
-    val pm10: Int,
+    val pm10: Double,
     val pm2_5: Double,
     val so2: Double,
     @Json(name = "us-epa-index")val usEpaIndex: Int
